Use AssertJ to rewrite TestDataMatchers
